Publish source files in package (#155)
With the latest version of react-scripts the configuration tries to
build source maps for the build. Because this package doesn't include
the source files, it generates warnings during the CRA build process.
The fix is to simply include the "src" directory in the package. It
slightly increases the package size, but not by enough to matter.